partial imguismo implementation

+++ Beetle/src/Beetle/Scene/SceneSerializer.cpp
@@ -15,7 +15,7 @@ namespace YAML
 	template <>
 	struct convert<glm::vec3>
 	{
-		static Node encode(const glm::vec3& rhs)
+		static Node encode(const glm::vec3 &rhs)
 		{
 			Node node;
 			node.push_back(rhs.x);
@@ -25,7 +25,7 @@ namespace YAML
 			return node;
 		}
 
-		static bool decode(const Node& node, glm::vec3& rhs)
+		static bool decode(const Node &node, glm::vec3 &rhs)
 		{
 			if (!node.IsSequence() || node.size() != 3)
 				return false;
@@ -40,7 +40,7 @@ namespace YAML
 	template <>
 	struct convert<glm::vec4>
 	{
-		static Node encode(const glm::vec4& rhs)
+		static Node encode(const glm::vec4 &rhs)
 		{
 			Node node;
 			node.push_back(rhs.x);
@@ -51,7 +51,7 @@ namespace YAML
 			return node;
 		}
 
-		static bool decode(const Node& node, glm::vec4& rhs)
+		static bool decode(const Node &node, glm::vec4 &rhs)
 		{
 			if (!node.IsSequence() || node.size() != 4)
 				return false;
@@ -68,26 +68,26 @@ namespace YAML
 namespace Beetle
 {
 
-	YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec3& v)
+	YAML::Emitter &operator<<(YAML::Emitter &out, const glm::vec3 &v)
 	{
 		out << YAML::Flow;
 		out << YAML::BeginSeq << v.x << v.y << v.z << YAML::EndSeq;
 		return out;
 	}
 
-	YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec4& v)
+	YAML::Emitter &operator<<(YAML::Emitter &out, const glm::vec4 &v)
 	{
 		out << YAML::Flow;
 		out << YAML::BeginSeq << v.x << v.y << v.z << v.w << YAML::EndSeq;
 		return out;
 	}
 
-	SceneSerializer::SceneSerializer(const Ref<Scene>& scene)
+	SceneSerializer::SceneSerializer(const Ref<Scene> &scene)
 		: m_Scene(scene)
 	{
 	}
 
-	static void SerializeEntity(YAML::Emitter& out, Entity entity)
+	static void SerializeEntity(YAML::Emitter &out, Entity entity)
 	{
 		out << YAML::BeginMap; // Entity
 		out << YAML::Key << "Entity" << YAML::Value << static_cast<uint32_t>(entity);
@@ -97,7 +97,7 @@ namespace Beetle
 			out << YAML::Key << "TagComponent";
 			out << YAML::BeginMap; // TagComponent
 
-			auto& tag = entity.GetComponent<TagComponent>().Tag;
+			auto &tag = entity.GetComponent<TagComponent>().Tag;
 			out << YAML::Key << "Tag" << YAML::Value << tag;
 
 			out << YAML::EndMap; // TagComponent
@@ -108,7 +108,7 @@ namespace Beetle
 			out << YAML::Key << "TransformComponent";
 			out << YAML::BeginMap; // TransformComponent
 
-			auto& tc = entity.GetComponent<TransformComponent>();
+			auto &tc = entity.GetComponent<TransformComponent>();
 			out << YAML::Key << "Translation" << YAML::Value << tc.Translation;
 			out << YAML::Key << "Rotation" << YAML::Value << tc.Rotation;
 			out << YAML::Key << "Scale" << YAML::Value << tc.Scale;
@@ -121,8 +121,8 @@ namespace Beetle
 			out << YAML::Key << "CameraComponent";
 			out << YAML::BeginMap; // CameraComponent
 
-			auto& cameraComponent = entity.GetComponent<CameraComponent>();
-			auto& camera = cameraComponent.Camera;
+			auto &cameraComponent = entity.GetComponent<CameraComponent>();
+			auto &camera = cameraComponent.Camera;
 
 			out << YAML::Key << "Camera" << YAML::Value;
 			out << YAML::BeginMap; // Camera
@@ -146,7 +146,7 @@ namespace Beetle
 			out << YAML::Key << "SpriteRendererComponent";
 			out << YAML::BeginMap; // SpriteRendererComponent
 
-			auto& spriteRendererComponent = entity.GetComponent<SpriteRendererComponent>();
+			auto &spriteRendererComponent = entity.GetComponent<SpriteRendererComponent>();
 			out << YAML::Key << "Color" << YAML::Value << spriteRendererComponent.Color;
 
 			out << YAML::EndMap; // SpriteRendererComponent
@@ -155,7 +155,7 @@ namespace Beetle
 		out << YAML::EndMap; // Entity
 	}
 
-	void SceneSerializer::Serialize(const std::string& filepath)
+	void SceneSerializer::Serialize(const std::string &filepath)
 	{
 		YAML::Emitter out;
 		out << YAML::BeginMap;
@@ -163,7 +163,7 @@ namespace Beetle
 		out << YAML::Key << "Entities" << YAML::Value << YAML::BeginSeq;
 		for (auto entId : m_Scene->m_Registry.view<TransformComponent>())
 		{
-			Entity entity = { entId, m_Scene.get() };
+			Entity entity = {entId, m_Scene.get()};
 			if (!entity)
 				continue;
 
@@ -176,13 +176,13 @@ namespace Beetle
 		fout << out.c_str();
 	}
 
-	void SceneSerializer::SerializeRuntime(const std::string& filepath)
+	void SceneSerializer::SerializeRuntime(const std::string &filepath)
 	{
 		// Not implemented
 		BT_CORE_ASSERT(false);
 	}
 
-	bool SceneSerializer::Deserialize(const std::string& filepath)
+	bool SceneSerializer::Deserialize(const std::string &filepath)
 	{
 		YAML::Node data = YAML::LoadFile(filepath);
 		if (!data["Scene"])
@@ -196,7 +196,7 @@ namespace Beetle
 		{
 			for (auto entity : entities)
 			{
-				uint32_t entityID = entity["Entity"].as<uint32_t>();
+				uint64_t entityID = entity["Entity"].as<uint64_t>();
 
 				std::string name = "Entity";
 				auto tagComponent = entity["TagComponent"];
@@ -211,7 +211,7 @@ namespace Beetle
 				if (transformComponent)
 				{
 					// Entities always have transforms
-					auto& tc = deserializedEntity.GetComponent<TransformComponent>();
+					auto &tc = deserializedEntity.GetComponent<TransformComponent>();
 					tc.Translation = transformComponent["Translation"].as<glm::vec3>();
 					tc.Rotation = transformComponent["Rotation"].as<glm::vec3>();
 					tc.Scale = transformComponent["Scale"].as<glm::vec3>();
@@ -220,9 +220,9 @@ namespace Beetle
 				auto cameraComponent = entity["CameraComponent"];
 				if (cameraComponent)
 				{
-					auto& cc = deserializedEntity.AddComponent<CameraComponent>();
+					auto &cc = deserializedEntity.AddComponent<CameraComponent>();
 
-					auto& cameraProps = cameraComponent["Camera"];
+					auto &cameraProps = cameraComponent["Camera"];
 					cc.Camera.SetProjectionType((SceneCamera::ProjectionType)cameraProps["ProjectionType"].as<int>());
 
 					cc.Camera.SetPerspectiveVerticalFOV(cameraProps["PerspectiveFOV"].as<float>());
@@ -240,7 +240,7 @@ namespace Beetle
 				auto spriteRendererComponent = entity["SpriteRendererComponent"];
 				if (spriteRendererComponent)
 				{
-					auto& src = deserializedEntity.AddComponent<SpriteRendererComponent>();
+					auto &src = deserializedEntity.AddComponent<SpriteRendererComponent>();
 					src.Color = spriteRendererComponent["Color"].as<glm::vec4>();
 				}
 			}
@@ -249,7 +249,7 @@ namespace Beetle
 		return true;
 	}
 
-	bool SceneSerializer::DeserializeRuntime(const std::string& filepath)
+	bool SceneSerializer::DeserializeRuntime(const std::string &filepath)
 	{
 		// Not implemented
 		BT_CORE_ASSERT(false);
